requirements.txt
numpy
humanizer
asyncio
aiohttp[speedups]
platformdirs
rich
pillow
